+++ src/EnemyStandard.h
@@ -25,7 +25,7 @@ class EnemyStandard final : public Object {
 public:
     EnemyStandard(std::string name, int energy, int score, 
                   sdlc::Surface& s, sdlc::Surface& hit_s,
-                  int init_y_vel);
+                  float init_y_vel);
     virtual ~EnemyStandard() = default;
 };
